+++ rocket_league/rocket_league.py
@@ -21,7 +21,7 @@ from harness_utils.output import (
 from harness_utils.process import terminate_processes
 from harness_utils.keras_service import KerasService
 from harness_utils.artifacts import ArtifactManager, ArtifactType
-from harness_utils.misc import LTTGamePadDS4
+from harness_utils.misc import LTTGamePadDS4, find_eg_game_version
 
 SCRIPT_DIRECTORY = os.path.dirname(os.path.realpath(__file__))
 LOG_DIRECTORY = os.path.join(SCRIPT_DIRECTORY, "run")
@@ -30,6 +30,7 @@ CONFIG_PATH = Path(f"C:\\Users\\{USERNAME}\\Documents\\My Games\\Rocket League\\
 PROCESS_NAME = "rocketleague.exe"
 EXECUTABLE_PATH = find_epic_executable()
 GAME_ID = "9773aa1aa54f4f7b80e44bef04986cea%3A530145df28a24424923f5828cc9031a1%3ASugar?action=launch&silent=true"
+gamefoldername = "rocketleague"
 am = ArtifactManager(LOG_DIRECTORY)
 gamepad = LTTGamePadDS4()
 
@@ -238,7 +239,7 @@ try:
         "resolution": format_resolution(width, height),
         "start_time": seconds_to_milliseconds(start_time),
         "end_time": seconds_to_milliseconds(end_time),
-        "game_version": find_game_version()
+        "game_version": find_eg_game_version(gamefoldername)
     }
 
     write_report_json(LOG_DIRECTORY, "report.json", report)